migrated to fast api but need to work on bicep
